# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.DefaultResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
...
  teacher.TeacherRequest:
    properties:
      address_detail:
        type: string
      address_region:
        type: string
      birth_date:
        description: YYYY-MM-DD
        type: string
      birth_place:
        type: string
      full_name:
        type: string
      gender:
        enum:
        - L
        - P
        type: string
      nip:
        type: string
      phone:
        type: string
    required:
    - full_name
    - gender
    - nip
    type: object
  teacher.TeacherResponse:
    properties:
      address_detail:
        type: string
      address_region:
        type: string
      birth_date:
        description: YYYY-MM-DD
        type: string
      birth_place:
        type: string
      created_at:
        type: string
      full_name:
        type: string
      gender:
        type: string
      id:
        type: string
      nip:
        type: string
      phone:
        type: string
      photo:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Documentation for LinguaLoop E-Learning Platform.
  title: LinguaLoop API
  version: "1.0"
paths:
...
  /teachers:
    get:
      description: Get a list of all teachers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/teacher.TeacherResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Get all teachers
      tags:
      - Teachers
    post:
      consumes:
      - application/json
      description: Add a new teacher to the system and auto-generate user account
      parameters:
      - description: Teacher Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/teacher.TeacherRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/teacher.TeacherResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Create new teacher
      tags:
      - Teachers
  /teachers/{id}:
    delete:
      description: Soft delete a teacher by ID
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Delete teacher
      tags:
      - Teachers
    get:
      description: Get detailed information of a teacher by their ID
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/teacher.TeacherResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Get teacher by ID
      tags:
      - Teachers
    put:
      consumes:
      - application/json
      description: Update existing teacher information
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: string
      - description: Teacher Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/teacher.TeacherRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/teacher.TeacherResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Update teacher
      tags:
      - Teachers
  /teachers/{id}/status:
    patch:
      description: Change teacher status between ACTIVE and INACTIVE
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/teacher.TeacherResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Toggle teacher status
      tags:
      - Teachers
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"